+++ app/condos/schema.py
@@ -97,19 +97,73 @@ class CreateBlock(graphene.Mutation):
     """Mutation from graphene for creating block"""
 
     number = graphene.String()
-
     class Arguments:
         number = graphene.String()
+
     @superuser_required
     def mutate(self, info, number):
-        block = Block(number=number)
+        block = Block(number=number, number_id=number)
         block.save()
 
         return CreateBlock(
             number=block.number)
 
+class DeleteApartment(graphene.Mutation):
+    apartment_number = graphene.Int()
+
+    class Arguments:
+        apartment_number = graphene.Int(required=True)
+
+    @superuser_required
+    def mutate(self, info, apartment_number):
+        apartment = Apartment.objects.get(number=apartment_number)
+        apartment.delete()
+
+class DeleteBlock(graphene.Mutation):
+    block_number = graphene.String()
+
+    class Arguments:
+        block_number = graphene.String(required=True)
+
+    def mutate(self, info, block_number):
+        block = Block.objects.get(number=block_number)
+        block.delete()
+
+class UpdateBlock(graphene.Mutation):
+    block = graphene.Field(BlockType)
+    number = graphene.String()
+    block_number = graphene.String()
+
+    class Arguments:
+        number = graphene.String(required=True)
+        block_number = graphene.String(required=True)
+
+    def mutate(self, info, number, block_number):
+        block = Block.objects.get(number=block_number)
+        block.number = number
+        block.save()
+        return UpdateBlock(block=block)
+
+class UpdateApartment(graphene.Mutation):
+    apartment = graphene.Field(ApartmentType)
+    number = graphene.String()
+    apartment_number = graphene.String()
+
+    class Arguments:
+        number = graphene.String(required=True)
+        apartment_number = graphene.String(required=True)
+
+    def mutate(self, info, number, apartment_number):
+        apartment = Apartment.objects.get(number=apartment_number)
+        apartment.number = number
+        apartment.save()
+        return UpdateApartment(apartment=apartment)
 
 
 class Mutation(graphene.ObjectType):
     create_apartment = CreateApartment.Field()
     create_block = CreateBlock.Field()
+    delete_block = DeleteBlock.Field()
+    delete_apartment = DeleteApartment.Field()
+    update_block = UpdateBlock.Field()
+    update_apartment = UpdateApartment.Field()